IT'S WORKING

+++ AVL-tree/functions.hpp
@@ -14,7 +14,7 @@ template <typename T>
 bool containsValues(std::deque<T*>& deque)
 {
     bool x = false;
-    for (int i = 0; i < deque.size(); i++)
+    for (size_t i = 0; i < deque.size(); i++)
     {
         x = x || deque[i] != nullptr;
         if (x)
